# schema/openapi.yaml
openapi: 3.1.0
info:
  title: api
  description: ""
  license:
    name: ""
  version: 0.1.0
paths:
  /auth/email/send-code:
    post:
      operationId: send_code
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeInput'
        required: true
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationCodeResponse'
        "400":
          description: invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "429":
          description: too many request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/email/verify-code:
    post:
      operationId: verify_code
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeInput'
        required: true
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          description: invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: unauthorized error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/google:
    post:
      operationId: auth_google
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthGoogleInput'
        required: true
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: unauthorized error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      operationId: auth_refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshInput'
        required: true
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: unauthorized error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    AuthGoogleInput:
      type: object
      required:
        - id_token
        - device_info
      properties:
        device_info:
          $ref: '#/components/schemas/DeviceInfoInput'
        id_token:
          type: string
    AuthRefreshInput:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    AuthResponse:
      type: object
      required:
        - user
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
    DeviceInfoInput:
      type: object
      required:
        - os
      properties:
        device_name:
          type:
            - string
            - "null"
        model_name:
          type:
            - string
            - "null"
        os:
          type: string
    ErrorCode:
      type: string
      enum:
        - INVALID_EMAIL
        - INVALID_DISPLAY_NAME
        - INVALID_VERIFICATION_CODE
        - INVALID_ACCESS_TOKEN
        - INVALID_REFRESH_TOKEN
        - RATE_LIMIT_EXCEEDED
        - SESSION_CREATION_FAILED
        - INTERNAL_ERROR
        - UNAUTHORIZED
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
    SendCodeInput:
      type: object
      required:
        - email
      properties:
        email:
          type: string
    UserResponse:
      type: object
      required:
        - id
        - display_name
        - role
      properties:
        display_name:
          type: string
        id:
          type: integer
          format: int64
        role:
          type: string
    VerificationCodeResponse:
      type: object
      required:
        - resend_cooldown_seconds
      properties:
        resend_cooldown_seconds:
          type: integer
          format: int32
          minimum: 0
    VerifyCodeInput:
      type: object
      required:
        - email
        - code
        - device_info
      properties:
        code:
          type: string
        device_info:
          $ref: '#/components/schemas/DeviceInfoInput'
        email:
          type: string